+++ changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+# v0.1.1
+
+
+## New Features
+
+* Added an option to rebuild road layout from current zone layout (graph view only)
+
+## Bugfixes
+
+* Crossroads are now displayed correctly in the road editor.
+
+---
+
 # v0.1.0 - Significant Changes